Update ExceptionOccured.php
Since the content property is public, isn't the "with" unnecessary? 
According to the documentation: https://laravel.com/docs/5.4/mail#view-data 
If a property is public it is automatically available in the view. 

Or am I missing something? 

Please let me know if I got this wrong since I am just learning all of this.

Thanks.